Final in-game framework complete
Going to bed now!

Added a Constants file, please store all of your globals there (makes
the code much tidier).

Now there is a little grey ball (a proxy for the jumping man which will
replace it when the sprite is done) that can jump up on mouse pressed
and drop otherwise.

There is also a buildings to jump over (again to be replaced with images
once complete).

None of the collision code is done yet.

If you change goingToWork bool you will get the going home version of
the game.