Create random_pickUps_template
Code that makes a random choice between drawing a green circle (placeholder for coffee), a red circle (sandwich) or nothing every three seconds. I've just drawn them as ellipses with position variables, but I would prefer to have them each as an object of a class. I haven't used classes in processing yet, so I'll look into it. This was mostly just a template for figuring out a method to generate the random objects; I may start from scratch on a new version at some point.